Update RedisIO to typescript

+++ SQL/stockml_exchange_install.sql
@@ -9,11 +9,11 @@ DROP TABLE IF EXISTS `candlestick_def`;
 
 CREATE TABLE `candlestick_def` (
   `time` bigint(20) NOT NULL DEFAULT '0',
-  `open` float NOT NULL DEFAULT '0',
-  `high` float NOT NULL DEFAULT '0',
-  `low` float NOT NULL DEFAULT '0',
-  `close` float NOT NULL DEFAULT '0',
-  `volume` float NOT NULL DEFAULT '0',
+  `open` double NOT NULL DEFAULT '0',
+  `high` double NOT NULL DEFAULT '0',
+  `low` double NOT NULL DEFAULT '0',
+  `close` double NOT NULL DEFAULT '0',
+  `volume` double NOT NULL DEFAULT '0',
   UNIQUE KEY `time` (`time`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
